git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@2587 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/db/postgres/PGType.java
@@ -19,9 +19,9 @@ public class PGType
 {
 	String rawType;
 	String formattedType;
-	int oid;
+	long oid;
 
-	public PGType(String raw, String formatted, int id)
+	public PGType(String raw, String formatted, long id)
 	{
 		if (raw.equals("int2"))
 		{
@@ -46,5 +46,5 @@ public class PGType
 		formattedType = formatted;
 		oid = id;
 	}
-	
+
 }
